Rename container to components
