+++ superduperdb/components/vector_index.py
@@ -5,9 +5,9 @@ import typing as t
 from overrides import override
 
 import superduperdb as s
-from superduperdb.container.component import Component
+from superduperdb.components.component import Component
 from superduperdb.base.document import Document
-from superduperdb.container.listener import Listener
+from superduperdb.components.listener import Listener
 from superduperdb.base.db import DB
 from superduperdb.misc.special_dicts import MongoStyleDict
 from superduperdb.vector_search.base import VectorIndexMeasureType
