+++ superduperdb/ext/cohere/model.py
@@ -5,9 +5,9 @@ import cohere
 import tqdm
 from cohere.error import CohereAPIError, CohereConnectionError
 
-from superduperdb.container.component import Component
-from superduperdb.container.encoder import Encoder
-from superduperdb.container.model import PredictMixin
+from superduperdb.components.component import Component
+from superduperdb.components.encoder import Encoder
+from superduperdb.components.model import PredictMixin
 from superduperdb.ext.utils import format_prompt, get_key
 from superduperdb.ext.vector.encoder import vector
 from superduperdb.misc.retry import Retry
