+++ superduperdb/ext/numpy/array.py
@@ -2,7 +2,7 @@ import typing as t
 
 import numpy
 
-from superduperdb.container.encoder import Encoder
+from superduperdb.components.encoder import Encoder
 from superduperdb.ext.utils import str_shape
