+++ superduperdb/ext/sentence_transformer/model.py
@@ -1,4 +1,4 @@
-from superduperdb.container.model import Model
+from superduperdb.components.model import Model
 
 
 class SentenceTransformer(Model):
