+++ superduperdb/ext/sklearn/model.py
@@ -5,8 +5,8 @@ import numpy
 from tqdm import tqdm
 
 from superduperdb.base.artifact import Artifact
-from superduperdb.container.metric import Metric
-from superduperdb.container.model import Model, _TrainingConfiguration
+from superduperdb.components.metric import Metric
+from superduperdb.components.model import Model, _TrainingConfiguration
 from superduperdb.base.db import DB
 from superduperdb.db.base.query import Select
 from superduperdb.db.query_dataset import QueryDataset
