+++ superduperdb/ext/torch/tensor.py
@@ -3,7 +3,7 @@ import typing as t
 import numpy
 import torch
 
-from superduperdb.container.encoder import Encoder
+from superduperdb.components.encoder import Encoder
 from superduperdb.ext.utils import str_shape
