+++ test/integration/conftest.py
@@ -14,8 +14,8 @@ except ImportError:
 
 from superduperdb import CFG
 from superduperdb.base.document import Document
-from superduperdb.container.listener import Listener
-from superduperdb.container.vector_index import VectorIndex
+from superduperdb.components.listener import Listener
+from superduperdb.components.vector_index import VectorIndex
 from superduperdb.db.mongodb.query import Collection
 from superduperdb.server.dask_client import dask_client
