+++ test/unittest/db/ibis/test_query.py
@@ -1,4 +1,4 @@
-from superduperdb.container.schema import Schema
+from superduperdb.components.schema import Schema
 from superduperdb.base.serializable import Serializable
 from superduperdb.db.ibis.field_types import dtype
 from superduperdb.db.ibis.query import IbisTable
