+++ test/unittest/model/test_transformers.py
@@ -6,7 +6,7 @@ try:
 except ImportError:
     torch = None
 
-from superduperdb.container.dataset import Dataset
+from superduperdb.components.dataset import Dataset
 from superduperdb.base.document import Document as D
 from superduperdb.db.mongodb.query import Collection
 from superduperdb.ext.transformers.model import (
